Rename book.py to search.py
