+++ opac/queries/book/__init__.py
@@ -1,2 +1,2 @@
-from .book import *  # noqa: F401 F403
+from .search import *  # noqa: F401 F403
 from .stocks import *  # noqa: F401 F403
